# Optimize method HttpStatusClass.valueOf(int)