--- a/codec-http/src/main/java/io/netty/handler/codec/http/HttpStatusClass.java
+++ b/codec-http/src/main/java/io/netty/handler/codec/http/HttpStatusClass.java
@@ -52,24 +52,24 @@ public enum HttpStatusClass {
         }
     };
 
-    private static final HttpStatusClass[] statusArray = new HttpStatusClass[5];
-
-    static {
-        statusArray[0] = INFORMATIONAL;
-        statusArray[1] = SUCCESS;
-        statusArray[2] = REDIRECTION;
-        statusArray[3] = CLIENT_ERROR;
-        statusArray[4] = SERVER_ERROR;
-    }
-
     /**
      * Returns the class of the specified HTTP status code.
      */
     public static HttpStatusClass valueOf(int code) {
-        if (UNKNOWN.contains(code)) {
-            return UNKNOWN;
+        switch (code / 100) {
+            // 1xx
+            case 1: return INFORMATIONAL;
+            // 2xx
+            case 2: return SUCCESS;
+            // 3xx
+            case 3: return REDIRECTION;
+            // 4xx
+            case 4: return CLIENT_ERROR;
+            // 5xx
+            case 5: return SERVER_ERROR;
+            // others
+            default: return UNKNOWN;
         }
-        return statusArray[(code / 100) - 1];
     }
 
     /**