--- a/microbench/src/main/java/io/netty/handler/codec/http/HttpStatusValueOfBenchmark.java
+++ b/microbench/src/main/java/io/netty/handler/codec/http/HttpStatusValueOfBenchmark.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2019 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
 package io.netty.handler.codec.http;
 import io.netty.microbench.util.AbstractMicrobenchmark;
 import io.netty.util.internal.SuppressJava6Requirement;
@@ -42,5 +57,4 @@ public class HttpStatusValueOfBenchmark extends AbstractMicrobenchmark {
         }
         return result;
     }
-
 }