--- a/microbench/src/main/java/io/netty/handler/codec/http/HttpStatusValueOfBenchmark.java
+++ b/microbench/src/main/java/io/netty/handler/codec/http/HttpStatusValueOfBenchmark.java
@@ -61,11 +61,12 @@ public class HttpStatusValueOfBenchmark extends AbstractMicrobenchmark {
         final SplittableRandom random = new SplittableRandom();
         // Equal the branch predictor.
         int equalDistributedArraySize = 16000;
-        int[] dataSwitchCase = new int[equalDistributedArraySize];
-        initDistributedData("dataSwitchCase", dataSwitchCase, random, 0.166, 0.166, 0.166,
-                0.166, 0.166, 0.166, 0.0);
+        int[] dataArrayIndex = new int[equalDistributedArraySize];
+        initDistributedData("dataArrayIndex", dataArrayIndex, random,
+                0.166, 0.166, 0.166, 0.166,
+                0.166, 0.166, 0.0);
         for (int i = 0; i < equalDistributedArraySize; i++) {
-            HttpStatusClass rs = HttpStatusClass.valueOf(dataSwitchCase[i]);
+            HttpStatusClass rs = HttpStatusClass.valueOf(dataArrayIndex[i]);
             bh.consume(rs);
         }
         data = new int[size];