Motivation:

The method `HttpStatusClass.valueOf(int)` can be optimized by replacing 'if branches' with 'switch cases'.
The 'switch cases' can be optimized by JIT to run faster than 'if branches'.
Refer https://stackoverflow.com/a/98024.

Modification:

Replace 'if-branches' with 'switch case'.

Result:

Remove the 'if-branches'.