@laosijikaichele 2023-08-11
> Integer division is pretty slow, while branches are often predictable.
> 
> I don't think this will be a beneficial change, unless you have a benchmark showing that it's worthwhile.

OK，will do a benchmark to check.