@laosijikaichele 2023-08-12
Did a preliminary benchmark, the results with 'switch-case':
```
Benchmark                           (size)   Mode  Cnt   Score   Error   Units
HttpStatusValueOfBenchmark.ofValue      25  thrpt   40  13.558 ± 0.576  ops/us
HttpStatusValueOfBenchmark.ofValue      32  thrpt   40  11.040 ± 0.076  ops/us
HttpStatusValueOfBenchmark.ofValue      63  thrpt   40   5.540 ± 0.065  ops/us
HttpStatusValueOfBenchmark.ofValue      95  thrpt   40   3.676 ± 0.043  ops/us
```
Before:
```
Benchmark                           (size)   Mode  Cnt  Score   Error   Units
HttpStatusValueOfBenchmark.ofValue      25  thrpt   40  9.835 ± 0.214  ops/us
HttpStatusValueOfBenchmark.ofValue      32  thrpt   40  5.896 ± 1.576  ops/us
HttpStatusValueOfBenchmark.ofValue      63  thrpt   40  4.059 ± 0.060  ops/us
HttpStatusValueOfBenchmark.ofValue      95  thrpt   40  2.584 ± 0.025  ops/us
```
Seems `switch-case` doing faster than `if-else` before.

Besides, the newest [commit](https://github.com/netty/netty/pull/13543/commits/8060704228ad8fff9ad300319724833fe5ace99a) replace the 'switch-case' with 'array-index-lookup', the benchmark results as following:
```
Benchmark                           (size)   Mode  Cnt   Score   Error   Units
HttpStatusValueOfBenchmark.ofValue      25  thrpt   40  15.787 ± 0.193  ops/us
HttpStatusValueOfBenchmark.ofValue      32  thrpt   40  12.238 ± 0.126  ops/us
HttpStatusValueOfBenchmark.ofValue      63  thrpt   40   6.270 ± 0.050  ops/us
HttpStatusValueOfBenchmark.ofValue      95  thrpt   40   4.197 ± 0.038  ops/us
```
It seems 'array-index-lookup' lookup gains more performance.

With VM:
```
openjdk version "17.0.8" 2023-07-18
OpenJDK Runtime Environment (build 17.0.8+7-Ubuntu-120.04.2)
OpenJDK 64-Bit Server VM (build 17.0.8+7-Ubuntu-120.04.2, mixed mode, sharing)
```
and 

`AMD EPYC 7763 64-Core Processor`

This is a very initial benchmark, any suggestion will be welcome.