@franz1981 2023-08-13
Yep @langchristian96 the main thing is:
- I didn't verified if switch cases branches are JIT profiled not, so worth trying to "pullute" the profile by making on SetUp +10000 benchmark iterations (look at my pr benchmark for the confuseJIT Param true) which can arrange the branches in different ways
- look at the SWAR indexOf benchmark trying to create a parametric long enough sequence of random inputs (use SplittableRandom with the same seed to get repeatable samples) to be fed into the benchmark cases, to see from s branch predictor perspective which one would work better. 


Last but not least:
- profile with perfasm
- profile with async profiler 
- gather numbers with perfnorm and a Linux box

If using a Numa platform please run the full benchmark with numactl --localalloc -N 0 and disable turbo frequency (you can use the latency profile with tuned from RHEL If you wish).

A number isn't meaningful if you don't know how/why has happened, IMHO.
The typical mistake is to assume averything to be inlined: verify it, it can produce dramatic difference in perf!