@laosijikaichele 2023-08-15
Been busy recently... Here is the latest update so far:

> Note: given that 100 is a constant from JIT perspective maybe there is some math trick (eg reciprocal multiplication) aiming to replace the division with something more appropriate, given that is stored into an int. The perfasm analysis should be done to verify it

@franz1981 You are right, the JIT did some constant divisor optimizations. 
Please see the following assembly of method `HttpStatusClass.valueOf(int)` which uses `code / 100`:

<img width="991" alt="c2-1" src="https://github.com/netty/netty/assets/29225782/ce86bcdf-de47-4fcc-a35b-b53bdfc80ece">

The quoted two lines above does the optimization of `code / 100`:

```
  0x00007f068908efb2:   imul   $0x51eb851f,%r10,%r10
  0x00007f068908efb9:   sar    $0x25,%r10
```
It can be translated to Java code like:
```
int fast_div100(int code) {
        // 1374389535L hex is 0x51eb851f
        return (int)(((code * 1374389535L) >> 37) - (code >> 31))
    }
```

Which means: `code / 100` can be optimized to `((code * 1374389535L) >> 37) - (code >> 31)`. 
If `code >= 0`, then it can be optimized to `(code * 1374389535L) >> 37`.

And the GCC did the same optimization: https://gcc.godbolt.org/z/3EqjGdfoW.

Will keep updating when further tests done. :)